requirements.txt
requests==2.32.3
pytest==8.3.3
python-dotenv==1.0.0
